+++ apps/decodex/src/orchestrator/retained_review_orchestration.rs
@@ -1,4 +1,4 @@
-struct RetainedReviewLane {
+pub(crate) struct RetainedReviewLane {
 	snapshot: PostReviewLaneSnapshot,
 	review_state: PullRequestReviewState,
 	orchestration_marker: ReviewOrchestrationMarker,
